Refactor: clean up redundant nullability and update clipboard handling
Remove unnecessary null-safe calls, non-null assertions, and redundant checks across `AppUpdateManager`, `ContentLoader` implementations, and unit tests.

Key changes:
- **Clipboard API Migration**: Update `LibraryScreen` to use the modern `LocalClipboard` API and `getClipEntry()` within a `LaunchedEffect` for reactive clipboard access.
- **WebView Hardening**: Remove deprecated `allowFileAccessFromFileURLs` and `allowUniversalAccessFromFileURLs` settings, which are no longer necessary for modern WebView security configurations.
- **Gradle Configuration**: Modernize `build.gradle.kts` by migrating source set definitions to use `directories.add` and adding the `-Xannotation-default-target=param-property` Kotlin compiler argument.
- **UI & Logic**: Simplify chapter text formatting in `LibraryItemCard` and remove redundant `.toInt()` conversions in `LibraryViewModelTest`.

Verified: `./gradlew test` (All tests passed, including `AsuraScansSourceTest` and `WebViewUtilsTest`).

+++ app/src/test/java/io/aatricks/easyreader/ui/viewmodel/LibraryViewModelTest.kt
@@ -264,7 +264,7 @@ class LibraryViewModelTest {
         assertEquals(baseTitle, insertedBaseTitle.firstValue)
         assertEquals(baseNovelUrl, insertedBaseNovelUrl.firstValue)
         assertEquals(sourceName, insertedSourceName.firstValue)
-        assertEquals(details.chapters.size, insertedTotalChapters.firstValue.toInt())
+        assertEquals(details.chapters.size, insertedTotalChapters.firstValue)
         verify(contentRepository, never()).prefetch(any(), any())
         assertEquals(latestUrl, loadedUrl)
         assertEquals(createdItem.id, loadedId)
